% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample data means.R
\docType{data}
\name{months.means}
\alias{months.means}
\title{Sample data means}
\format{
A vector of length 12
}
\usage{
months.means
}
\description{
The average proportion of customers that stayed on the website
for longer than 5 minutes each month.
}
\keyword{datasets}
