% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample data standard deviations.R
\docType{data}
\name{months.sd}
\alias{months.sd}
\title{Sample data standard deviations}
\format{
A vector of length 12
}
\usage{
months.sd
}
\description{
The standard deviations of the proportion of customers that stayed on the website
for longer than 5 minutes each month.
}
\keyword{datasets}
