% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{one_rep_profit}
\alias{one_rep_profit}
\title{Utility function used in function 'profit_nn_sim()' to simulate one set of potential outcomes
Draws a true mean for each arm and generates N observations from each arm
Returns profits and error rates under perfect information, test & roll, and Thomson sampling}
\usage{
one_rep_profit(n, N, s, mu, sigma, K, TS = FALSE)
}
\arguments{
\item{n}{vector of length 2 containing the sample sizes}

\item{N}{the deployment population}

\item{s}{vector of length 2 containing the standard deviations of the outcome}

\item{mu}{vector of length 2 containing the means of the prior on the mean response}

\item{sigma}{vector of length 2 containing the standard deviations of the prior on the mean response}

\item{K}{the number of arms (treatments)}

\item{TS}{whether or not to run Thomson sampling, 'TRUE' or 'FALSE'}
}
\value{
The profits and error rates under perfect information, test & roll, and Thomson sampling
}
\description{
Utility function used in function 'profit_nn_sim()' to simulate one set of potential outcomes
Draws a true mean for each arm and generates N observations from each arm
Returns profits and error rates under perfect information, test & roll, and Thomson sampling
}
