% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{plot_prior_mean_resp_nn}
\alias{plot_prior_mean_resp_nn}
\title{Plots prior densities against mean response (profit per customer)}
\usage{
plot_prior_mean_resp_nn(mu, sigma)
}
\arguments{
\item{mu}{means of the prior on the mean response}

\item{sigma}{standard deviations of the prior on the mean response}
}
\value{
graph plotting prior density vs. profit per customer
}
\description{
Plots prior densities against mean response (profit per customer)
}
\examples{
#plot_prior_mean_resp_nn(mu=c(5,10), sigma=c(10,10))
}
