% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{plot_prior_resp_nn}
\alias{plot_prior_resp_nn}
\title{Plots prior densities against response (profit per customer)}
\usage{
plot_prior_resp_nn(s, mu, sigma)
}
\arguments{
\item{s}{known standard deviations of the response}

\item{mu}{means of the prior on the mean response}

\item{sigma}{standard deviations of the prior on the mean response}
}
\value{
graph plotting prior density vs. profit per customer
}
\description{
Plots prior densities against response (profit per customer)
}
\examples{
plot_prior_resp_nn(s=c(10,20), mu=c(5,10), sigma=c(10,10))
}
