% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{profit_nn}
\alias{profit_nn}
\title{Computes the per-customer profit for test & roll with 2 arms}
\usage{
profit_nn(n, N, s, mu, sigma, log_n = FALSE)
}
\arguments{
\item{n}{vector of length 2 containing the sample sizes}

\item{N}{size of deployment population}

\item{s}{vector of length 2 containing the known standard deviations of the outcome}

\item{mu}{vector of length 2 containing the means of the prior on the mean response}

\item{sigma}{vector of length 2 containing the standard deviations of the prior on the mean response}

\item{log_n}{whether or not log(n) is an input rather than n (to avoid negative solutions), 'TRUE' or 'FALSE'}
}
\value{
per-customer profit for N customers
}
\description{
Computes the per-customer profit for test & roll with 2 arms
}
\examples{
profit_nn(n=100, N=10000, s=.1, mu=c(.7,.5), sigma=c(.2,.2))
profit_nn(n=c(100,200), N=10000, s=.1, mu=c(.7,.5), sigma=c(.2,.2))
}
