% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{profit_nn_sim}
\alias{profit_nn_sim}
\title{Computes the per-customer profit for test & roll with K arms}
\usage{
profit_nn_sim(n, N, s, mu, sigma, K = 2, TS = FALSE, R = 1000)
}
\arguments{
\item{n}{sample sizes for test & roll}

\item{N}{deployment population}

\item{s}{standard deviations of the outcome}

\item{mu}{means of the priors on the mean response (vector of length 1 or K)}

\item{sigma}{standard deviations of the priors on the mean response (vector of length 1 or K)}

\item{K}{number of arms}

\item{TS}{whether or not to run Thomson sampling, 'TRUE' or 'FALSE'}

\item{R}{number of simulation replications}
}
\value{
A list containing the profit, regret, and error rates
}
\description{
Computes the per-customer profit for test & roll with K arms
}
\examples{
profit_nn_sim(n=100, N=1000, s=10, mu=10, sigma=10, K=2, TS=FALSE, R=100)
profit_nn_sim(n=c(100,200,300), N=1000, s=c(.1,.2,.3), mu=c(.1,.2,.3), sigma=c(.01,.03,.05), K=3, TS=FALSE, R=100)

}
