% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{profit_perfect_nn}
\alias{profit_perfect_nn}
\title{Computes per-customer profit with perfect information
Where response is normal with symmetric normal priors}
\usage{
profit_perfect_nn(mu, sigma)
}
\arguments{
\item{mu}{means of the prior on the mean response}

\item{sigma}{standard deviations of the prior on the mean response}
}
\value{
per-customer profit with perfect information
}
\description{
Computes per-customer profit with perfect information
Where response is normal with symmetric normal priors
}
\examples{
profit_perfect_nn(mu=.7, sigma=.02)
}
