% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{test_size_nn}
\alias{test_size_nn}
\title{Computes the profit-maximizing test size for test & roll with 2 arms
(asymmetric case not yet implemented)}
\usage{
test_size_nn(N, s, mu, sigma)
}
\arguments{
\item{N}{size of deployment population}

\item{s}{vector of length 2 containing the standard deviations of the outcome}

\item{mu}{vector of length 2 containing the means of the prior on the mean response}

\item{sigma}{}
}
\value{
a vector containing the sample sizes
}
\description{
Computes the profit-maximizing test size for test & roll with 2 arms
(asymmetric case not yet implemented)
}
\examples{
test_size_nn(N=10000, s=.1, mu=c(.7,.7), sigma=c(.05,.05))
test_size_nn(N=10000, s=c(.1,.2), mu=c(.7,.7), sigma=c(.05,.05))
}
