% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{test_size_nn_sim}
\alias{test_size_nn_sim}
\title{Computes the profit-maximizing test size for a multi-armed test & roll}
\usage{
test_size_nn_sim(N, s, mu, sigma, K = 2, R = 1000)
}
\arguments{
\item{N}{deployment population}

\item{s}{standard deviations of the response (length 1(symmetric) or K)}

\item{mu}{vector of length K containing means of the priors on the mean response}

\item{sigma}{vector of length K containing the standard deviations of the priors on the mean response}

\item{K}{number of arms (treatments)}

\item{R}{number of simulation replications}
}
\value{
a list with the sample sizes and expected profit per customer
}
\description{
Computes the profit-maximizing test size for a multi-armed test & roll
}
\examples{
test_size_nn_sim(N=1000, s=.1, mu=.1, sigma=.05, K=2, R=1000)
test_size_nn_sim(N=1000, s=c(.1,.2,.3), m=c(.1,.2,.3), sigma=c(.05,.08,.1), K=3, R=1000)
}
