% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{error_rate_nn}
\alias{error_rate_nn}
\title{Computes the rate of incorrect deployments
Where response is normal with symmetric normal priors}
\usage{
error_rate_nn(n, s, sigma)
}
\arguments{
\item{n}{vector of length 2 containing the sample sizes}

\item{s}{vector of length 2 containing the standard deviations of the outcome}

\item{sigma}{vector of length 2 containing the standard deviations of the prior on the mean response}
}
\value{
rate of incorrect deployments
}
\description{
Computes the rate of incorrect deployments
Where response is normal with symmetric normal priors
}
