% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{one_rep_test_size}
\alias{one_rep_test_size}
\title{Utility function used in test_size_nn_sim() to simulate one set of potential outcomes
Returns profits for all possible equal sample sizes}
\usage{
one_rep_test_size(n_vals, N, s, mu, sigma, K)
}
\arguments{
\item{N}{deployment population}

\item{s}{standard deviations of the outcome}

\item{mu}{means of the priors on the mean response}

\item{sigma}{standard deviations of the priors on the mean response}

\item{K}{number of arms (treatments)}
}
\value{

}
\description{
Utility function used in test_size_nn_sim() to simulate one set of potential outcomes
Returns profits for all possible equal sample sizes
}
