% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{test_eval_nn}
\alias{test_eval_nn}
\title{Provides summary of a test & roll plan}
\usage{
test_eval_nn(n, N, s, mu, sigma)
}
\arguments{
\item{sigma}{}
}
\value{
a data frame containing summary statistics such as profit per customer,
profits from test phase, error rates, etc.
}
\description{
Provides summary of a test & roll plan
}
