% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{test_size_nn_sim}
\alias{test_size_nn_sim}
\title{Computes the profit-maximizing test size for a multi-armed test & roll}
\usage{
test_size_nn_sim(N, s, mu, sigma, K = 2, R = 1000)
}
\arguments{
\item{N}{deployment population}

\item{s}{standard deviations of the response (length 1(symmetric) or K)}

\item{mu}{vector of length K containing means of the prior on the mean response}

\item{sigma}{vector of length K containing the standard deviations of the priors on the mean response}

\item{K}{number of arms (treatments)}

\item{R}{number of simulation replications}
}
\value{

}
\description{
Computes the profit-maximizing test size for a multi-armed test & roll
}
